!!! TASK: Make configuration non editable
Previously it was possible to change the
configuration in the backend module.
This didn’t make a lot of sense, because
deployments and using configuration in
a `Settings.yaml` file overwrote this config.

The best practise is to have the configuration
in the site package or add it during
deployment.