FEATURE: #19 Add caching for api requests
With this change node statistics are cached
by default based on their „period“ argument.

1 year periods: 1 day caching
1-x days: 1 hour caching

This can speed up the inspector statistics views
quite a bit in a multi editor environment.

This change is also a preparation to support
more detailed analysis of pages in the future.
For example having custom views to check
a pages performance.

Resolves: #19